Phase 1: Completed Smart Contracts and Testnet Deployment

+++ contracts/MediRegistry.py
@@ -1,5 +1,4 @@
-from algopy import ARC4Contract, itxn, String, arc4, Global
-
+from algopy import ARC4Contract, itxn, String, arc4, Global, UInt64
 class MediRegistry(ARC4Contract):
     @arc4.abimethod
     def mint_medicine_batch(self, batch_name: String, ipfs_metadata_url: String) -> arc4.UInt64:
@@ -7,8 +6,8 @@ class MediRegistry(ARC4Contract):
             asset_name=batch_name,
             unit_name=String("MED"),
             url=ipfs_metadata_url,
-            total=100,
-            decimals=0,
+            total=UInt64(100),
+            decimals=UInt64(0),
             manager=Global.current_application_address,
             freeze=Global.current_application_address,
         ).submit()
